Add Combination of User and Students

+++ src/Learnify.Application/Students/IStudentAppService.cs
@@ -21,9 +21,9 @@ namespace Learnify.Students
 
         Task<Student> CreateAsync(Student student);
 
-        Task<Student?> UpdateAsync(int id, Student student);
+        Task<Student?> UpdateAsync(long id, Student student);
 
-        Task<Student?> DeleteAsync(int id);
+        Task<Student?> DeleteAsync(long id);
 
         Task<bool> ExistingEnrollment(int StudentId, int CourseId);
     }
